# apps/web/src/shared/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fleet Maintenance Decision API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/demo/catalog": {
      "get": {
        "tags": [
          "customer trials"
        ],
        "summary": "Trial Catalog",
        "operationId": "trial_catalog_api_demo_catalog_get",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialCatalogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/trials": {
      "get": {
        "tags": [
          "customer trials"
        ],
        "summary": "Trial List",
        "operationId": "trial_list_api_demo_trials_get",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrialResponse"
                  },
                  "title": "Response Trial List Api Demo Trials Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customer trials"
        ],
        "summary": "Trial Create",
        "operationId": "trial_create_api_demo_trials_post",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/trials/{trial_id}": {
      "get": {
        "tags": [
          "customer trials"
        ],
        "summary": "Trial Detail",
        "operationId": "trial_detail_api_demo_trials__trial_id__get",
        "parameters": [
          {
            "name": "trial_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trial Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/trials/{trial_id}/planning": {
      "post": {
        "tags": [
          "customer trials"
        ],
        "summary": "Trial Plan",
        "operationId": "trial_plan_api_demo_trials__trial_id__planning_post",
        "parameters": [
          {
            "name": "trial_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trial Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/trials/{trial_id}/jobs": {
      "get": {
        "tags": [
          "customer trials"
        ],
        "summary": "Trial Jobs",
        "operationId": "trial_jobs_api_demo_trials__trial_id__jobs_get",
        "parameters": [
          {
            "name": "trial_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trial Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response Trial Jobs Api Demo Trials  Trial Id  Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/trials/{trial_id}/history": {
      "get": {
        "tags": [
          "customer trials"
        ],
        "summary": "Trial History",
        "operationId": "trial_history_api_demo_trials__trial_id__history_get",
        "parameters": [
          {
            "name": "trial_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trial Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryImport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts",
        "operationId": "list_alerts_api_alerts_get",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertResponse"
                  },
                  "title": "Response List Alerts Api Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/{alert_id}/acknowledgements": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_api_alerts__alert_id__acknowledgements_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/history/{component_id}": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "History",
        "operationId": "history_api_alerts_history__component_id__get",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Component Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertResponse"
                  },
                  "title": "Response History Api Alerts History  Component Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertAcknowledgementResponse": {
        "properties": {
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "actor",
          "created_at"
        ],
        "title": "AlertAcknowledgementResponse"
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "component_id": {
            "type": "string",
            "title": "Component Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Id"
          },
          "acknowledged_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged By"
          },
          "acknowledgements": {
            "items": {
              "$ref": "#/components/schemas/AlertAcknowledgementResponse"
            },
            "type": "array",
            "title": "Acknowledgements"
          },
          "episode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "policy_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Policy Context"
          }
        },
        "type": "object",
        "required": [
          "id",
          "component_id",
          "state",
          "reason",
          "policy_version",
          "assessment_id",
          "acknowledged_by",
          "acknowledgements"
        ],
        "title": "AlertResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryImport": {
        "properties": {
          "source_version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "title": "Source Version"
          },
          "engine_identity": {
            "type": "string",
            "maxLength": 120,
            "pattern": "^NASA_CMAPSS:FD001:(train|test):[1-9][0-9]*$",
            "title": "Engine Identity"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/HistoryRow"
            },
            "type": "array",
            "maxItems": 10000,
            "minItems": 1,
            "title": "Rows"
          },
          "previous_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_version",
          "engine_identity",
          "rows"
        ],
        "title": "HistoryImport"
      },
      "HistoryRow": {
        "properties": {
          "cycle": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Cycle"
          },
          "values": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "maxItems": 24,
            "minItems": 24,
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cycle",
          "values"
        ],
        "title": "HistoryRow"
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "attempt": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "state",
          "attempt",
          "result"
        ],
        "title": "JobResponse"
      },
...
      "TrialCatalogResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "history": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistoryImport"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Sha256"
          }
        },
        "type": "object",
        "required": [
          "available",
          "reason",
          "model_id",
          "history",
          "source_sha256"
        ],
        "title": "TrialCatalogResponse"
      },
      "TrialRequest": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^trial-[a-f0-9]{32}$",
            "title": "Id"
          },
          "aircraft_label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Aircraft Label"
          },
          "cutoff_cycle": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Cutoff Cycle"
          },
          "cycles_per_day": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Cycles Per Day"
          },
          "duration_slots": {
            "type": "integer",
            "maximum": 14.0,
            "minimum": 1.0,
            "title": "Duration Slots"
          },
          "deadline_slot": {
            "type": "integer",
            "maximum": 14.0,
            "minimum": 1.0,
            "title": "Deadline Slot"
          },
          "spare_on_hand": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Spare On Hand"
          },
          "arrival_slot": {
            "type": "integer",
            "maximum": 13.0,
            "minimum": 0.0,
            "title": "Arrival Slot"
          },
          "history": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistoryImport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "aircraft_label",
          "cutoff_cycle",
          "cycles_per_day",
          "duration_slots",
          "deadline_slot",
          "spare_on_hand",
          "arrival_slot"
        ],
        "title": "TrialRequest"
      },
      "TrialResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "aircraft_label": {
            "type": "string",
            "title": "Aircraft Label"
          },
          "component_id": {
            "type": "string",
            "title": "Component Id"
          },
          "part_id": {
            "type": "string",
            "title": "Part Id"
          },
          "import_id": {
            "type": "string",
            "title": "Import Id"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "cutoff_cycle": {
            "type": "integer",
            "title": "Cutoff Cycle"
          },
          "cycles_per_day": {
            "type": "number",
            "title": "Cycles Per Day"
          },
          "duration_slots": {
            "type": "integer",
            "title": "Duration Slots"
          },
          "deadline_slot": {
            "type": "integer",
            "title": "Deadline Slot"
          },
          "spare_on_hand": {
            "type": "integer",
            "title": "Spare On Hand"
          },
          "arrival_slot": {
            "type": "integer",
            "title": "Arrival Slot"
          },
          "arrival_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Id"
          },
          "baseline_scenario_id": {
            "type": "string",
            "title": "Baseline Scenario Id"
          },
          "supply_scenario_id": {
            "type": "string",
            "title": "Supply Scenario Id"
          },
          "engine_identity": {
            "type": "string",
            "title": "Engine Identity"
          },
          "history_sha256": {
            "type": "string",
            "title": "History Sha256"
          },
          "history_origin": {
            "type": "string",
            "title": "History Origin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "aircraft_label",
          "component_id",
          "part_id",
          "import_id",
          "model_id",
          "cutoff_cycle",
          "cycles_per_day",
          "duration_slots",
          "deadline_slot",
          "spare_on_hand",
          "arrival_slot",
          "arrival_id",
          "baseline_scenario_id",
          "supply_scenario_id",
          "engine_identity",
          "history_sha256",
          "history_origin"
        ],
        "title": "TrialResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}